Moved gauss_integration and find_eshelby function to helper header
* Can now be used in external unit test
